LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-011-expected.txt


PASS Should use the last fallback position initially
FAIL Should use the third fallback position with enough space above assert_equals: Anchored element should be at the top of anchor expected 235 but got 435
FAIL Should use the second fallback position with enough space right assert_equals: Anchored element should be at the right of anchor expected 358 but got 158
PASS Should use the first fallback position with enough space above and right
